Refactor bindings, no more monolithic files
One downside of this change is that building the C++ bindings now
takes longer, since it now builds multiple .o files. But this makes
the code easier to navigate and less error-prone, since each file can
import freely from namespaces without causing collisions with each
other.
